init 140
init 140

init 140

init 140

init 140

initial push

140 initial push

+++ comm/mail/test/browser/shared-modules/OpenPGPTestUtils.sys.mjs
@@ -188,13 +188,12 @@ export const OpenPGPTestUtils = {
       localPassphraseProvider,
       passphrase != null && keepPassphrase,
       pgpBlock,
-      false,
       []
     );
 
     if (!result || result.exitCode !== 0) {
       throw new Error(
-        `EnigmailKeyRing.importKey failed with result "${result.errorMsg}"!`
+        `RNP.importSecKeyBlockImpl() failed with result "${result.errorMsg}"!`
       );
     }
     if (!result.importedKeys || !result.importedKeys.length) {
@@ -215,15 +214,16 @@ export const OpenPGPTestUtils = {
    * @param {nsIWindow} parent - The parent window.
    * @param {nsIFile} file - A valid file containing an OpenPGP key.
    * @param {boolean} [isBinary] - false for ASCII armored files
+   * @param {string} [acceptance=null] - The acceptance to set for the imported key.
    * @returns {Promise<string[]>} - A list of ids for the key(s) imported.
    */
-  async importKey(parent, file, isBinary) {
+  async importKey(parent, file, isBinary, acceptance = null) {
     const data = await IOUtils.read(file.path);
     const txt = lazy.MailStringUtils.uint8ArrayToByteString(data);
     const errorObj = {};
     const fingerPrintObj = {};
 
-    const result = lazy.EnigmailKeyRing.importKey(
+    const result = await lazy.EnigmailKeyRing.importKeyAsync(
       parent,
       false,
       txt,
@@ -232,12 +232,13 @@ export const OpenPGPTestUtils = {
       errorObj,
       fingerPrintObj,
       false,
-      []
+      [],
+      acceptance
     );
 
     if (result !== 0) {
       console.debug(
-        `EnigmailKeyRing.importKey failed with result "${result}"!`
+        `EnigmailKeyRing.importKeyAsync failed with result "${result}"!`
       );
       return [];
     }
@@ -247,12 +248,12 @@ export const OpenPGPTestUtils = {
   /**
    * Updates the acceptance value of the provided key(s) in the database.
    *
-   * @param {string|string[]} id - The id or list of ids to update.
+   * @param {string|string[]} idOrList - The id or list of ids to update.
    * @param {string} acceptance - The new acceptance level for the key id.
    * @returns {string[]} - A list of the key ids processed.
    */
-  async updateKeyIdAcceptance(id, acceptance) {
-    const ids = Array.isArray(id) ? id : [id];
+  async updateKeyIdAcceptance(idOrList, acceptance) {
+    const ids = Array.isArray(idOrList) ? idOrList : [idOrList];
     for (const id of ids) {
       const key = lazy.EnigmailKeyRing.getKeyById(id);
       const email = lazy.EnigmailFuncs.getEmailFromUserID(key.userId);
@@ -270,13 +271,16 @@ export const OpenPGPTestUtils = {
    * Removes a key by its id, clearing its acceptance and refreshing the
    * cache.
    *
-   * @param {string|string[]} id - The id or list of ids to remove.
+   * @param {string|string[]} idOrList - The id or list of ids to remove.
    * @param {boolean} [deleteSecret=false] - If true, secret keys will be removed too.
    */
-  async removeKeyById(id, deleteSecret = false) {
-    const ids = Array.isArray(id) ? id : [id];
+  async removeKeyById(idOrList, deleteSecret = false) {
+    const ids = Array.isArray(idOrList) ? idOrList : [idOrList];
     for (const id of ids) {
       const key = lazy.EnigmailKeyRing.getKeyById(id);
+      if (!key) {
+        throw new Error(`Could not find key by id=${id}`);
+      }
       await lazy.RNP.deleteKey(key.fpr, deleteSecret);
       await lazy.PgpSqliteDb2.deleteAcceptance(key.fpr);
     }
